fix: fix old tests and add new test file
